Add proof-of-concept script to download Sitka rainfall data
This is a node script that can be run from the command line (with
`node --experimental-modules rainfall.js`) that fetches observed rainfall
data from the MesoWest API and predicted future rainfall amounts from the
NWS API. Currently it just prints the results to the console.

Notes/features:
- This includes two different ways of getting past rainfall. I implemented
the "all observations for the past 24 hours" one first, then realized that
was both more than we need and a bit limited because it seems to return
'null' for both no-rain and no-observation cases. So I made a new function
that uses the 'precip' endpoint to get only the past 3 hours, but I didn't
remove the original yet.
- I got the parameters of the model and wrote them into a formula so this
can return risk rating in addition to rainfall amounts.
- The MesoWest API supports returning results in the local timezone, but
NWS is UTC-only, so I added the "luxon" package and made a helper function
to ensure all timestamps are returned in local Sitka time.
- I decided to use Axios for the requests, because:
  - I've used it before
  - It has a fairly nice API (sensible invocation syntax, full async support)
  - It supports both Node and browser environments, with the exact same API,
    so if we decided this should actually go in the front-end, it would be
    easy to move it there.
  - Since the plan is not to put it in the front-end, package size is not
    a big concern.