Add a 3hr look-back to risk calculation
Rather than just calculate risk for each period, we want to let a period
of high rainfall affect the risk for the period that comes after. So this
adds a 3hr look-back so that the risk calculation will be based on the higher
of the total for the current period or the previous one.
For the nearest prediction, whose previous forecast would be in the past, it
uses the observed 3hr rainfall.

This also removes the getPast24hrRainfall function, which used the MesoWest
'timeseries' endpoint, because it's not useful. The precip endpoint is more
reliable for our purposes and takes parameters that let us get exactly the
totals we need.