Add precip values in inches
We only want to use inches on the front-end. We could convert as soon as
we download the numbers, but then we have to think about how much precision
to preserve for the model vs. how much to show to users. Just using
millimeters for the calculations and including inches only for display
is simpler.