Simplify getStaticProps and add commented-out log for debugging
Now that the rainfall script is called within the build, it's not as easy
to see exactly what it's doing. So having a console.log of the whole thing
that we can just uncomment seems handy. Also this switches to just passing
along the whole object rather than destructuring it into separate variables
then building again exactly how it was.