Add multiplier to exaggerate risk for testing
Dangerously-high rainfall is very rare, but we want to be able to see what
the site would look like if it happened.  One simple way to make that happen
is to multiply the actual rainfall by some factor that will put it up into
medium- or high-risk range.  This adds a constant to do that.  Set to 1
normally, but easy to change when debugging.