Make sure days are listed in chronological order
The "days" list for the 3-day forecast is built up as an object keyed
by numerical day-of-week, so when it gets written to an array it always
comes out in weekday order, even if the weeks included should be ordered
differently (e.g. because it's Saturday, so Monday should come 3rd).
This adds a timestamp field and sorts the array by it.