Write rainfall.json to filesystem in index.js
It turns out it's kind of tricky to get the rainfall/risk data into the
detail page without calling the function (and thus the API) a whole bunch
of times. This is one way to do it--write it to the filesystem in index.js,
so it'll be there to just load via import in [timestamp].js.

I'm not sure this will work in production, because there's no explicit
dependency between the pages so the detail page could get built with
missing or stale data. But for development, when stale data isn't a big
concern, it will work.