Add short timestamp field for use in detail URLs
Dropping the crufty bit from the end of the timestamps when using them
as detail page URLs seems good, but doing the same split-and-index in
three different places across two files is not so good. So this adds a
field to the data that the pages can just use directly.

The detail view file is called [timestamp].js, so the param passed from
getStaticPaths has to be called 'timestamp'. This is slightly confusing
now that it's actually using the 'shortTimestamp' property, but it's
fairly self-contained, so it doesn't seem worth renaming the file over.