Scale numeric risk rating based on low/med/high ranges
The linear low/med/high scale looks weird when we just the modeled
probability of a landslide on it, because the breakpoints for what percentage
probability translates to each risk level are 1% and 70%. So especially
the medium-risk points will often show up below the middle of the scale
if we just plot probability.
This changes the `riskProb` value to scale the numerical rating differently
within each range, so for low-risk predictions it shows where the value falls
between 0 and 1%, for medium-risk it shows where it falls between 1% and 70%,
etc.  I reused the `riskProb` field name rather than making a new one because
we only use it in two places, and we want the new behavior in both those
places, and I couldn't really think of a better name anyway.

Resolves #19