Limit historical rainfall to 10 years, and store in inches
Showing historical rainfall going back to 2002 doesn't make sense when
it was actually only coded to show landslides for the later portion.
This limits the historical range to 10 years so it won't look like we're
claiming there weren't any landslides in the aughts.
This also converts the file to having values stored in inches, since that's
what we want in the frontend and it doesn't make sense to do the conversion
every time.

Resolves #29