Add weather alert lookup to data fetching script (#34)
There was already code related to weather alerts in the frontend and the
rainfall script, but it was hard-coded to inactive. This makes it call the
NWS current alerts API, using the location of the Sitka airport.

Per https://www.weather.gov/media/documentation/docs/NWS_Geolocation.pdf,
using a point request (rather than a county or a zone) seems like the way
to get all alerts that apply to that location, regardless of what level
(zone or county) they're defined for.

As noted in a comment, if this request fails it will just return
"active: false". It will still log an error, but it won't prevent the
script from succeeding if this is the observation and forecast endpoints
work.

Resolves #16