Add axios-retry to retry failed API requests (#37)
Adds the 'axios-retry' package and uses it in getRainfall.js to retry
API requests that fail in a way that could be transient, i.e. when there's
a 500-series error.  If all the retries fail, the last error comes through
and gets logged and either caught (for alerts) or not (for observations
and forecasts) as usual.

Resolves #31