Rename new more-detailed datetime label to 'dateTimeDetails'
Renames what was the datetime label and is now an object containing a few
different formatted versions of the datetime to make it more clear that
we're talking about that custom object, not an instance of the standard
DateTime class.

Proposed fixup to PR #44, coded up because the diff is big enough that
it would be cumbersome to do it as line-by-line suggestions.