Run updates on a fixed schedule
New rainfall observations are normally recorded at 53 minutes after the
hour, so always running a few minutes after that will slightly increase
the average freshness of the site's current risk estimate. So this changes
the schedule from "every 20 minutes", which doesn't make any particular
guarantees about when the events will fall, to an exact schedule that
still runs every 20 minutes, but one of the updates always happens at
57 minutes after the hour.

This also has the side benefit that it will be easier to tell from a
glance at the site whether its last update succeeded. With no fixed
schedule, you would only know for sure that an update failed if the site
is more than about 22 minutes old, but with a fixed schedule, you can look
at the timestamp at any time and know if the last update was successful.