Only use previous rainfall when risk is higher, and always show current
Changes the getRainfall script to only use the rainfall totals from a previous
period when they would actually cause the risk to be higher, and changes the
front end to always show the actual amount for the time period in question,
even if the risk level is based on an earlier amount.

This creates a field, riskIsElevatedFromPreviousPrecip, to identify when this
situation applies. It's used inside the script, and provided in the data JSON,
since the front end will need to use it to show a message explaining why the
risk and precip amount don't seem to line up, but for now it's not used by the
front end.