Use current, not buffered, precip amounts for text and graphs
Changes the detail view so that it uses the rainfall amount for the period
in question, not the max amount for the period or the two preceding it, for
the text and chart.  The risk is still based on the highest amount within
the lookback window, but the rainfall amount is the actual amount for the
period.