Send current and buffered risk levels to front end
This changes the getRainfall script to include both bufferedRiskLevel
(based on the highest precip over the current or past two 3-hour periods)
and currentRiskLevel (based on only the period in question), and incorporates
that into the front end. The name change only applies to the granular
data--the daily summaries and chart values still just call it 'riskLevel'.