Add indicator for when risk level is high from earlier rainfall
We need the front end to show a message when the risk level and the precip
for a period don't match up because the risk is based on earlier heavy rain.
This adds logic to the getRainfall script to figure that out, and a field
(a very wordy field, but unambiguous!) to the data JSON to let the front end
know the answer.