Fix rounding error in previous precip calculation
The value of `prevPrecip` was ending up as NaN because
- there was rain in the past 3 hours but not in the preceding 3, so the 3-hour
and 6-hour totals were the same
- the prevPrecip calculation was subtracting the rounded 3-hour total (0.03)
from the unrounded 6-hour total (0.029999999999972715) resulting in a very
small, negative amount (-2.7283730830163222e-14)
- The rounding function was using a trick involving converting to and from
scientific notation to get accurate decimal rounding, but it turns out that
fails badly when the incoming number is already in scientific notation.

So this changes both the subtraction and the rounding to prevent this
situation from coming up again. It now subtracts before rounding, which
should preserve accuracy but avoid any miniscule remainders, and the rounding
function now uses a different strategy that works fine on inputs that
already have an exponent.