Adjust build schedule to try to catch delayed rainfall readings
The rainfall readings normally happen at :53, but sometimes there are
extras, so we've been doing builds at :57 and then at 20-minute increments
after that. But lately the readings have been showing up late in the
API, so this pushes the :57 one back to :59 and adds another build at :08
(15 minutes after the reading should be taken) to hopefully catch delayed
values without waiting until :17.  But the :17 one is still there in case
it's delayed more than 15 minutes (and the :37 one is still there for the
reason it always was--to keep the "last updated" tag from getting too old
and to catch any extra readings taken during the hour, which sometimes
happens during periods of heavier rain.