Set http_version to http2and3 on Cloudfront distribution
This option was enabled manually and has worked fine, so this sets it in
the config so that Terraform won't try to roll it back.

The option was added in v4.27.0 of the AWS provider
(https://github.com/hashicorp/terraform-provider-aws/releases/tag/v4.27.0),
so this also bumps the required provider version to the latest release.
And it bumps the required Terraform version as well. Which might not be
necessary, but Bryan and I are both using v1.3, so we don't actually know
if older versions would work and we might as well avoid having to find out.