Switch to SSSC's Mapbox token and base layer style
Even though SSSC had set up a Mapbox account, the map was still using the
token and basemap style from Jeff's personal account, that he had put in
early on as a proof-of-concept. I copied the style into the science center's
account and this switches the code to use that layer and a token from
that account.