Fix compiler warnings from Resources.js
These have probably been here a while, but they just cause warnings in
the terminal when viewing the static resource pages, so they're not very
visible. But I saw them when looking at the Areas At Risk page, and the
warnings basically tell you how to fix them, so might as well.