Change the chart label for "Current risk" to "Past 3 hours" (#77)
The historical rainfall chart doesn't actually graph the current risk, it
graphs the most recent 3-hour rainfall. Which might be close enough, except
that the current risk could be affected by earlier rainfall, in which case
it could get confusing to see a low value on the chart labeled as a high
risk.

So this changes the chart label to "Past 3 hours".  Which also solves the
problem of "current" possibly being interpreted as "current forecast for
the next 3 hours".

It turns out the label gets set by `getRainfall.js`, since the labels for
other values are timestamps that get processed/formatted in that script, so
to reduce the amount of special-case handling in the front end, we just made
"Current risk" a datetime label.  But of the four places it was used in the
front end (chart, chart legend, window title, page title), only two should
change. So this adds a bit of logic to make `timestamp.js` use a `pageTitle`
prop that matches the datetime label for future periods but gets set to
"Current risk" for the current period.

Resolves #76