Update Synoptic token to new account
Synoptic agreed to let us use an "academic" account to get free access
to the endpoint we need.  We made a new account, sitkalandslide@element84.com.
There was talk of transferring the token we were already using, but that
didn't happen, and it's very easy to just make a new one and update the code.
So that's what this is doing.