Move AWS_PROFILE up and export it in cipublish
I think this script was working because it was always getting run after we had
already exported AWS_PROFILE to run the Terraform steps. In a new shell with no
AWS_PROFILE set, it was failing for lack of credentials.
The environment variables were already hard-coded in the script, and that's fine,
but this:
- moves them up to the top
- sets AWS_PROFILE to read from the environment and fall back to 'sitka', rather
  than always using 'sitka' regardless of what the environment says
- exports AWS_PROFILE, which is evidently necessary to make it available to the
  commands below