Copy rainfall.json to data/ in exported app
As we learned on 12/26 when the NWS API went down and there was no one around
with an already-built app to help deal with it, it's very hard to run this app
without a `rainfall.json` file, and very hard to get one unless both of the APIs
are working.
This is a super-simple way to address that problem, by uploading the
rainfall.json file for each build, so there will always be one on S3 that we can
use. It might also be useful for facilitating debugging--if the site does
something weird that seems data-related, we can grab the exact data file it was
built with, rather than having to figure out how to reproduce the issue from
new data.

Also:
- Updates `ci.yml` to run `yarn export` rather than `yarn run next export`.
It should have been doing that anyway, since that's what production does, and
now it will actually make a difference.
- Adds a README note to let people who might need the file know that it exists