Tag published image with git SHA
Updates cipublish to tag the published image with the git SHA in addition
to 'latest'.  It won't do anything functional, but it will make it easier
to know what we're looking at in ECR.

Also changes the environment variables used by cipublish to read from the
calling environment if there are values set, falling back to the defaults.