Save rainfall.json to archive bucket on build
Adds a command to the build process to copy the new rainfall.json file
into the archive bucket in addition to publishing the site.

Since the command to be run by the Dockerfile was getting even more complex,
and Docker was already mad that we were providing CMD in shell mode rather
than with a JSON list, I moved it into a new script, within `frontend/`.
So now the particulars of what the build does when it runs live in there,
and the Dockerfile just sets it as its entrypoint.

Note: I also added a 'mkdir -p out/data to the 'export' command in
package.json. The actual build has been working fine without it, but in local
testing I got an error that the directory didn't exist, and making sure it
does won't do any harm.