Add year to file path in archive bucket
The archived rainfall.json files get generated 4 times per hour (so 96/day,
2880/month, 35k/year), which will add up in the bucket. S3 can handle it,
but at a certain point it becomes annoying to work with because the file
list is so long. On the other hand too much directory structure might get
annoying when trying to investigate a time period that spans folders.

But it makes sense to at least group them by year, so this changes the
script to include the year in the file path.  I'll manually move the ones
that have already been written.

The good news is the files all have full timestamps, so the directory
structure is just for convenience, we don't lose information if we
move files around.