Add '--no-use' to prevent silent failure of activate_yvm
Sometimes my scripts would just fail without any output, which turned
out to be because `.nvm/nvm.sh` returns a nonzero exit status in some
cases and the script has '-e' set.
I'm still a little confused by why the script does that, but I found
the recommendation to add '--no-use' in an issue discussion and it
seems to do the trick for me.