Add Google Analytics, using nextjs-google-analytics
Adds Google Analytics to the site using the nextjs-google-analytics package.
Next.js maintains a package of their own that's very similar (called
@next/third-parties), but it requires newer versions of Node and React than
what we're using. I checked to see if upgrading would somehow be easy and not
break things, but the answer was 'no', so we're using this older, 3rd-party one.

The question of how to supply the analytics ID to the production site but not
have it active in development is a little tricky in the case of this app because
the production site is fully static and has no environment variables.
But the pay the package works, you can supply the ID as a parameter to the tag
OR as an environment variable, and the latter gets used if both are given.
So I resolved the dilemma by hard-coding the ID, which means it gets supplied
to the statically built production environment, then adding the environment
variable to the command that runs the dev server, so it will override the ID
with a fake value.

Resolves #91