Add a script for viewing historical rainfall in dev
Adds a Python utility script designed to make it easy to view the site as it
would have been under the conditions captured in archived rainfall.json files.

The script makes a list of the available archive files and lets you select which
one you want to see, then it just copies that file over 'data/rainfall.json',
taking advantage of the quick/automatic reloading features of the dev site to
show the site as it would have been and let you browse backwards and forwards
through time.

Related to issue #89