Store archived rainfall by month, and adjust viewer script
Turns out thousands of files in a single directory on S3 is probably too
much. Not that S3 itself will be bothered, but the web interface becomes
pretty unusable and seemingly even the CLI won't return all the results
when you list them.

So this changes the `publish` script that writes the archive files to
include the month in the file path, and changes the `historical_rainfall.py`
viewer script to accomodate the new format.

This also includes a few other tweaks to the script, including a command
to change years, now that there are two different years in the archive
data. Note on this: I decided to keep the index that the script uses by
year, not break it down by month, because it seemed like that would make
navigation more difficult to implement and not have any major advantages.