Revise/update development instructions in README
This makes some revisions to (hopefully) make the development instructions
a little more complete and easy to follow. Since running a dev site is part
of the process for viewing historical rainfall, it's not going to just be
E84 engineers and designers doing it soon.

It also changes the organization slightly, to put everything related to local
development inside the "Development" section. And it removes the generic Next.js
README from the `frontend` directory, in favor of just linking to nextjs.org
from the main README.