Remove hard coded MESOWEST_TOKEN
This was previously hard coded. Now this is read from the
environment. For devlopment, this is picked up from the .env
file, which is downloaded from S3 during `setup`. For deployments.
this is added to .tfvars which should be downloaded from S3
before running terraform commands. It is also added to the
ECS task's environment, so when new images are baked and
deployed, they have it already.

README is updated accordingly.